Revised port tests.
